Fix errors that arise from limited transaction history
Fix errors in normalize.fidelity.option.symbol and
load.schwab.transactions when there are only cash transactions in the
history.

Fix error in load.schwab.transactions when there are only small-sized
trades that incur no fees in the history.

Fix error in append.contra.tx when there are no buy or sell transactions
in the history.

Fix error in adjust.for.schwab.corporate.actions when there are no
reverse splits in the history.

Fix error in calc.portfolio.size.snapshot.with.day.over.day.gain when
there is only one day's worth of transactions in the history.

Fix inaccurate size snapshots and day over day gains in portfolio
summary when "future transactions" exist in the history. This happens on
days a BrokerageLink transfer is scheduled since the NetBenefit
account's history shows the cash outflow on T + 0 while the Fidelity
BrokerageLink account's history shows the cash inflow on T + 1.

Define the initial day over day gain as the initial day's gain in size
vs. time plots.