Add Dividends.Minus.Fees column to portfolio summary and do not adjust cash cost on involuntary actions
The cost of acquiring cash should only increase when we deposit money
into the account or sell a security. The cost of cash positions should
only decline when we withdraw money out of the account or buy a
security. Cash acquisitions from dividends are at no cost to us. Cash
losses from fees and commissions do not unintuitively decrease the cost
of acquiring cash.