Plot options payoff diagrams per expiration and download spot prices of options positions
Calculate knots of intrinsic value charts as well as the coordinates
where intrinsic value intersects with current cost, value, and spot
price. Connect the dots with lines.

Add get.current.price support for symbols on which we made no
transactions, e.g. for getting the spot price of an option when we never
traded the underlying. Download spot price history of options positions
in refresh.alphavantage.prices and degrade gracefully when we have no
price history due to Alpha Vantage failing to recognize the underlying
symbol.