Avoid unnecessary Alpha Vantage price refreshes, support Schwab options exercise transactions, support histories with no options transactions, and loosen mutual fund price check for Fidelity trades
Files in the Alpha Vantage directory containing just a date column with
the values 0001-01-01 and 9999-12-31 should be treated as sentinel
values that disable price refreshes for the symbol. Avoid overflowing
px.next.date in refresh.alphavantage.prices and append stub high, low,
and close columns with NA values in load.alphavantage.prices.

Do not truncate refreshed Alpha Vantage prices more tightly than the
dates in the existing CSV for the symbol. This avoids repeatedly
redownloading CSVs when refreshing prices for two accounts that have
trades for the same symbol but on different date ranges.

"Exchange or Exercise" has the same quantity sign as "Expired" and
"Assigned", so treat them the same.

Tolerate a $0.01 error in Fidelity mutual fund transaction price in
addition to the 0.001 error in quantity and correct quantity tolerance
direction for sales.

Skip wash journal transactions in recent.transaction.price.provider.